+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/component/ComponentService.scala
@@ -18,7 +18,7 @@ import pl.touk.nussknacker.ui.NotFoundError
 import pl.touk.nussknacker.ui.component.DefaultComponentService.{getComponentDoc, getComponentIcon}
 import pl.touk.nussknacker.ui.config.ComponentLinksConfigExtractor.ComponentLinksConfig
 import pl.touk.nussknacker.ui.process.processingtypedata.ProcessingTypeDataProvider
-import pl.touk.nussknacker.ui.process.{ProcessCategoryService, ProcessService}
+import pl.touk.nussknacker.ui.process.{ProcessCategoryService, ProcessService, UserCategoryService}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -36,7 +36,10 @@ object DefaultComponentService {
 
   def apply(
       componentLinksConfig: ComponentLinksConfig,
-      processingTypeDataProvider: ProcessingTypeDataProvider[ProcessingTypeData, ComponentIdProvider],
+      processingTypeDataProvider: ProcessingTypeDataProvider[
+        ProcessingTypeData,
+        (ComponentIdProvider, ProcessCategoryService)
+      ],
       processService: ProcessService,
       categoryService: ProcessCategoryService,
       additionalUIConfigProvider: AdditionalUIConfigProvider,
@@ -70,7 +73,10 @@ object DefaultComponentService {
 
 class DefaultComponentService private (
     componentLinksConfig: ComponentLinksConfig,
-    processingTypeDataProvider: ProcessingTypeDataProvider[ProcessingTypeData, ComponentIdProvider],
+    processingTypeDataProvider: ProcessingTypeDataProvider[
+      ProcessingTypeData,
+      (ComponentIdProvider, ProcessCategoryService)
+    ],
     processService: ProcessService,
     categoryService: ProcessCategoryService,
     additionalUIConfigProvider: AdditionalUIConfigProvider,
@@ -80,9 +86,8 @@ class DefaultComponentService private (
 
   import cats.syntax.traverse._
 
-  private val componentObjectsService = new ComponentObjectsService(categoryService)
-
-  private def componentIdProvider: ComponentIdProvider = processingTypeDataProvider.combined
+  private def componentIdProvider = processingTypeDataProvider.combined._1
+  private def categoryService     = processingTypeDataProvider.combined._2
 
   override def getComponentsList(implicit user: LoggedUser): Future[List[ComponentListElement]] = {
     for {
@@ -124,7 +129,8 @@ class DefaultComponentService private (
       processingType: ProcessingType,
       user: LoggedUser
   ): Future[List[ComponentListElement]] = {
-    val userCategories               = categoryService.getUserCategories(user)
+    val userCategoryService          = new UserCategoryService(categoryService)
+    val userCategories               = userCategoryService.getUserCategories(user)
     val processingTypeCategories     = categoryService.getProcessingTypeCategories(processingType)
     val userProcessingTypeCategories = userCategories.intersect(processingTypeCategories)
 
@@ -152,6 +158,7 @@ class DefaultComponentService private (
     processService
       .getFragmentsDetails(processingTypes = Some(List(processingType)))(user)
       .map { fragments =>
+        val componentObjectsService = new ComponentObjectsService(categoryService)
         // We assume that fragments have unique component ids ($processing-type-fragment-$name) thus we do not need to validate them.
         val componentObjects = componentObjectsService.prepare(
           processingType,