+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/ProcessToolbarService.scala
@@ -13,10 +13,10 @@ trait ProcessToolbarService {
   def getProcessToolbarSettings(process: BaseProcessDetails[_]): ProcessToolbarSettings
 }
 
-class ConfigProcessToolbarService(config: Config, categories: List[String]) extends ProcessToolbarService {
+class ConfigProcessToolbarService(config: Config, getCategories: () => List[String]) extends ProcessToolbarService {
 
   private val categoriesProcessToolbarConfig: Map[String, ProcessToolbarsConfig] =
-    categories
+    getCategories()
       .map(category => category -> ProcessToolbarsConfigProvider.create(config, Some(category)))
       .toMap