+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ComponentResourcesSpec.scala
@@ -33,7 +33,7 @@ class ComponentResourcesSpec
 
   private lazy val componentService = DefaultComponentService(
     ComponentLinksConfigExtractor.extract(testDbConfig),
-    testProcessingTypeDataProvider.mapCombined(_ => defaultComponentIdProvider),
+    testProcessingTypeDataProvider.mapCombined(_ => (defaultComponentIdProvider, processCategoryService)),
     processService,
     processCategoryService,
     TestAdditionalUIConfigProvider,